feat/#14: migration to TS helpers
