--- src/helpers/error.js
@@ -1,18 +1,0 @@
-class ErrorHandler extends Error {
-    constructor(statusCode, message) {
-      super(message);
-      this.statusCode = statusCode;
-    }
-}
-
-
-const handleError = (err, res) => {
-    const { statusCode, message } = err;
-    res.status(statusCode).json({
-      status: "error",
-      statusCode,
-      message
-    });
-};
-
-module.exports = {ErrorHandler, handleError}
