--- src/helpers/response.js
@@ -1,6 +1,0 @@
-const buildResponse = (res, status, result) => {
-    res.status(status);
-    res.json(result);
-};
-
-module.exports = {  buildResponse };
